requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
google-generativeai
